crawler/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
chardet==4.0.0
idna==3.3
multidict==5.2.0
pycodestyle==2.8.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.2
yarl==1.7.0
